+++ templates/listOfSps-tpl.php
@@ -4,122 +4,239 @@
  *
  * @author Dominik Baránek <0Baranek.dominik0@gmail.com>
  */
-
 $this->data['header'] = '';
 $this->data['head'] = '<link rel="stylesheet"  media="screen" type="text/css" href="' . SimpleSAML\Module::getModuleUrl('perun/res/css/listOfSps.css')  . '" />';
-
 $this->includeAtTemplateBase('includes/header.php');
 
-$attrNames = $this->data['attrNames'];
-$facilitiesWithAttributes = $this->data['facilitiesWithAttributes'];
+$statistics = $this->data['statistics'];
+$attributesToShow = $this->data['attributesToShow'];
+$samlServices = $this->data['samlServices'];
+$oidcServices = $this->data['oidcServices'];
 
-$columns = '[';
-$columns .= '{label: "Id", type: "number"},';
-$columns .= '{label: "Name", type: "string"},';
-$columns .= '{label: "Description", type: "string"},';
-if (!empty($attrNames)) {
-	$facilityAttributes = array_values($facilitiesWithAttributes)[0]['facilityAttributes'];
-	foreach ($attrNames as $attrName) {
-		if (typeIsSupported($facilityAttributes[$attrName]['type'])) {
-			$columns .= '{label: "' . $facilityAttributes[$attrName]['displayName'] . '", type: "';
-			if (strpos($facilityAttributes[$attrName]['type'], 'Integer')) {
-				$columns .= 'number';
-			} else if (strpos($facilityAttributes[$attrName]['type'], 'Boolean')) {
-				$columns .= 'boolean';
-			} else if (strpos($facilityAttributes[$attrName]['type'], 'String') || strpos($facilityAttributes[$attrName]['type'], 'Array')) {
-				$columns .= 'string';
-			}
-			$columns .= '"},';
-		}
-	}
-	$columns = substr($columns, 0, -1) . ']';
-} else {
-	$columns .= ']';
-}
+$allServices = array_merge($samlServices, $oidcServices);
+usort($allServices, 'sortByName');
 
-$rows = '[';
-foreach ($facilitiesWithAttributes as $facilityWithAttributes) {
-	$rows .= '{c:[';
-	$rows .= '{v: "' . $facilityWithAttributes['facility']->getId() . '"}, ';
-	$rows .= '{v: "' . $facilityWithAttributes['facility']->getName() . '"}, ';
-	$rows .= '{v: "' . $facilityWithAttributes['facility']->getDescription() . '"}, ';
-	foreach ($attrNames as $attrName) {
-		if (typeIsSupported($facilityWithAttributes['facilityAttributes'][$attrName]['type'])) {
-			if ((strpos($facilityWithAttributes['facilityAttributes'][$attrName]['type'], 'Array'))) {
-				$rows .= '{v: "';
-				foreach ($facilityWithAttributes['facilityAttributes'][$attrName]['value'] as $value) {
-					$rows .= $value . '; ';
-				}
-				if (!empty($facilityWithAttributes['facilityAttributes'][$attrName]['value'])) {
-					$rows = substr($rows, 0, -2) . '"}, ';
-				} else {
-					$rows .= '"}, ';
-				}
-			} else {
-				$rows .= '{v: "' . $facilityWithAttributes['facilityAttributes'][$attrName]['value'] . '"}, ';
-			}
-		}
-	}
-	$rows = substr($rows, 0, -2) . ']},';
-}
-if (!empty($facilitiesWithAttributes)) {
-	$rows = substr($rows, 0, -1) . ']';
-} else {
-	$rows .= ']';
-}
-
+$productionServicesCount = $statistics['samlServicesCount'] - $statistics['samlTestServicesCount'] + $statistics['oidcServicesCount'] - $statistics['oidcTestServicesCount'];
+$testServicesCount = $statistics['samlTestServicesCount'] + $statistics['oidcTestServicesCount'];
+$samlProductionCount = $statistics['samlServicesCount'] - $statistics['samlTestServicesCount'];
+$oidcProductionCount = $statistics['oidcServicesCount'] - $statistics['oidcTestServicesCount'];
 ?>
-	<html>
-	<head>
-		<script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-		<script type="text/javascript">
-			google.charts.load('current', {'packages':['corechart', 'controls']});
-			google.charts.setOnLoadCallback(drawDashboard);
 
-			function drawDashboard() {
-
-				var dashboard = new google.visualization.Dashboard(
-					document.getElementById('dashboard_div'));
-
-				var control = new google.visualization.ControlWrapper({
-					'controlType': 'StringFilter',
-					'containerId': 'stringFilter',
-					'options': {
-					    'matchType': 'any',
-						'filterColumnLabel': 'Name',
+<div class="row">
+    <div class="col-md-12">
+        <div class="row">
+            <div class="col-md-12">
+                <h3><?php echo $this->t('{perun:listOfSps:statistics}')?></h3>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-md-6">
+                <table class="table table-stats">
+                    <tr>
+                        <th></th>
+                        <th><?php echo $this->t('{perun:listOfSps:production_services}')?></th>
+                        <th><?php echo $this->t('{perun:listOfSps:test_services}')?></th>
+                    </tr>
+                    <tr>
+                        <th><?php echo $this->t('{perun:listOfSps:all}')?></th>
+                        <th><?php echo $productionServicesCount ?></th>
+                        <th><?php echo $testServicesCount ?></th>
+                    </tr>
+                    <tr>
+                        <th><?php echo $this->t('{perun:listOfSps:saml}')?></th>
+                        <th><?php echo $samlProductionCount ?></th>
+                        <th><?php echo $statistics['samlTestServicesCount'] ?></th>
+                    </tr>
+                    <tr>
+                        <th><?php echo $this->t('{perun:listOfSps:oidc}')?></th>
+                        <th><?php echo $oidcProductionCount ?></th>
+                        <th><?php echo $statistics['oidcTestServicesCount'] ?></th>
+                    </tr>
+                </table>
+            </div>
+            <div class="col-md-6 center">
+                <canvas id="myChart"></canvas>
+            </div>
+        </div>
+    </div>
+</div>
+<div class="row">
+    <div class="col-md-12">
+        <div class="row">
+            <div class="col-md-12">
+                <h3><?php echo $this->t('{perun:listOfSps:services}')?></h3>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-md-12 table-responsive">
+                <table class="table table-striped" id="table1">
+                    <thead>
+                    <tr>
+                        <th><?php echo $this->t('{perun:listOfSps:name}')?></th>
+                        <th><?php echo $this->t('{perun:listOfSps:authenticate_protocol}')?></th>
+                        <th><?php echo $this->t('{perun:listOfSps:description}')?></th>
+						<?php
+						foreach($attributesToShow as $attr) {
+							if (!is_null($samlServices) && !empty($samlServices)) {
+								echo "<th class='" . getClass(array_values($samlServices)[0]['facilityAttributes'][$attr]) . "'>" . array_values($samlServices)[0]['facilityAttributes'][$attr]['displayName'] . "</th>";
+							}
+						}
+						?>
+                    </tr>
+                    </thead>
+                    <tbody>
+					<?php
+					foreach($allServices as $service) {
+						if (is_null($service['showOnServiceList']) || is_null($service['showOnServiceList']['value'])
+							|| empty($service['showOnServiceList']['value']) || !($service['showOnServiceList']['value'])) {
+							continue;
+						}
+						echo "<tr>";
+						echo "<td>" . printServiceName($service) . "</td>";
+						if(array_key_exists($service["facility"]->getID(), $samlServices)) {
+							echo "<td>" . $this->t('{perun:listOfSps:saml}') . "</td>";
+						} else {
+							echo "<td>" . $this->t('{perun:listOfSps:oidc}') . "</td>";
+						}
+						echo "<td>" . $service['facility']->getDescription() . "</td>";
+						foreach ($attributesToShow as $attr) {
+							$value = printAttributeValue($service['facilityAttributes'][$attr], $service, $attr);
+							echo $value;
+						}
 					}
-				});
-
-				var chart  = new google.visualization.ChartWrapper({
-					'chartType': 'Table',
-					'containerId': 'table',
-				});
-
-				var data = new google.visualization.DataTable({
-					cols: <?php echo $columns ?>,
-					rows: <?php echo $rows ?>
-				});
-
-				dashboard.bind(control, chart);
-				dashboard.draw(data);
-			}
-		</script>
-	</head>
-	<body>
-		<h2><?php echo $this->t('{perun:perun:listOfSps_header}'); ?></h2>
-		<div id="listOfSps">
-			<div id="stringFilter"></div>
-			<div id="table"></div>
-		</div>
-	</body>
-	</html>
+					echo "</tr>";
+					?>
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    </div>
+</div>
 
 <?php
 $this->includeAtTemplateBase('includes/footer.php');
 
-function typeIsSupported($type) {
-	return strpos($type, 'Integer') ||
-		strpos($type, 'Boolean') ||
-		strpos($type, 'String') ||
-		strpos($type, 'Array');
+function printServiceName($service) {
+	if (is_null($service['loginURL']) || is_null($service['loginURL']['value']) || empty($service['loginURL']['value'])) {
+		return $service['facility']->getName();
+	} else {
+		return "<a class='customLink' href='" . $service['loginURL']['value'] . "'>" . $service['facility']->getName() . "</a>";
+	}
 }
+
+function printAttributeValue($attribute, $service, $attr) {
+	$value = $attribute['value'];
+	if((is_null($value) || empty($value)) && $attribute['type'] != "java.lang.Boolean") {
+		return "<td class='center'>&horbar;</td>";
+	}
+	$string = '';
+	if($attribute['type'] == "java.lang.String" || $attribute['type'] == "java.lang.LargeString") {
+		if(filter_var($value, FILTER_VALIDATE_URL)) {
+			$string = "<a class='customLink' href='" . $value . "'>" . $value . "</a>";
+		} else {
+			$string = $value;
+		}
+	} elseif($attribute['type'] == "java.lang.Integer") {
+		$string = $value;
+	} elseif($attribute['type'] == "java.lang.Boolean") {
+		if(!is_null($value) && $value) {
+			$string = "&#x2714;";
+		} else {
+			$string = "&#x2715;";
+		}
+	} elseif ($attribute['type'] == "java.util.ArrayList" || $attribute['type'] == "java.lang.LargeArrayList") {
+		$string = "<ul>";
+		foreach($value as $v) {
+			$string .= "<li>" . $v . "</li>";
+		}
+		$string .= "</ul>";
+	} elseif ($attribute['type'] == "java.util.LinkedHashMap") {
+		$string = "<ul>";
+		foreach($value as $k => $v) {
+			$string .= "<li>" . $k ." &rarr; " .  $v . "</li>";
+		}
+		$string .= "</ul>";
+	}
+	if(!empty($string)) {
+		return "<td class='".getClass($service['facilityAttributes'][$attr]) ."'>" . $string . "</td>";
+	} else {
+		return '<td/>';
+	}
+}
+
+function getClass($attribute) {
+	if($attribute['type'] == "java.lang.String") {
+		return "string";
+	} elseif($attribute['type'] == "java.lang.Integer") {
+		return "integer";
+	} elseif($attribute['type'] == "java.lang.Boolean") {
+		return "boolean";
+	} elseif ($attribute['type'] == "java.util.ArrayList" || $attribute['type'] == "java.util.LargeArrayList") {
+		return "array";
+	} elseif ($attribute['type'] == "java.util.LinkedHashMap") {
+		return "map";
+	} else {
+		return '';
+	}
+}
+
+function sortByName($a, $b) {
+	return strcmp(strtolower($a['facility']->getName()), strtolower($b['facility']->getName()));
+}
+?>
+
+<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.3/Chart.bundle.js"></script>
+
+<script>
+    var ctx = document.getElementById("myChart").getContext('2d');
+    var myChart = new Chart(ctx, {
+        type: 'bar',
+        data: {
+            labels: [<?php echo
+				'"' . $this->t('{perun:listOfSps:saml_production}') . '"' . ", " .
+				'"' . $this->t('{perun:listOfSps:saml_test}') . '"' . ", " .
+				'"' . $this->t('{perun:listOfSps:oidc_production}') . '"' . ", " .
+				'"' . $this->t('{perun:listOfSps:oidc_test}') . '"'
+				?>
+            ],
+            datasets: [{
+                label: '',
+                data: [<?php echo $samlProduction . ', ' . $statistics['samlTestServicesCount'] . ', ' . $oidcProduction . ', ' . $statistics['oidcTestServicesCount']?>],
+                backgroundColor: [
+                    'rgba(255, 99, 132, 0.2)',
+                    'rgba(54, 162, 235, 0.2)',
+                    'rgba(255, 206, 86, 0.2)',
+                    'rgba(75, 192, 192, 0.2)'
+                ],
+                borderColor: [
+                    'rgba(255,99,132,1)',
+                    'rgba(54, 162, 235, 1)',
+                    'rgba(255, 206, 86, 1)',
+                    'rgba(75, 192, 192, 1)'
+                ],
+                borderWidth: 1
+            }]
+        },
+        options: {
+            scales: {
+                yAxes: [{
+                    ticks: {
+                        beginAtZero: true,
+                        callback: function (value) { if (Number.isInteger(value)) { return value; } },
+                        stepSize: 1
+                    }
+                }]
+            },
+            legend: {
+                display: false
+            },
+            tooltips: {
+                callbacks: {
+                    label: function(tooltipItem) {
+                        return tooltipItem.yLabel;
+                    }
+                }
+            }
+        }
+    });
+</script>